refactor(audit): separate file sink internals
Split file_store.rs (2090 lines) into 3 modules, no behavior change:
- file_sink.rs: recovery/classification logic, private sync AuditSink mock trait, lock/open/seal helpers
- writer.rs: writer_loop, write_draft_to_chain, eviction sentinel
- store.rs: AuditFileStore handle + WriterMsg + all tests

Prepares for generalizing over the new public cosmian_kms_interfaces::AuditSink trait in the next commit.
