+++ crate/server/src/core/audit/mod.rs
@@ -1,3 +1,5 @@
-mod file_store;
+mod file_sink;
+mod store;
+mod writer;
 
-pub(crate) use file_store::AuditFileStore;
+pub(crate) use store::AuditFileStore;
